udpated with visibility

+++ linkSharing/src/main/java/com/ttn/linkSharing/repositories/ResourceRepository.java
@@ -3,5 +3,5 @@ package com.ttn.linkSharing.repositories;
 import com.ttn.linkSharing.entities.Resource;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ResourceRepository extends CrudRepository<Resource, Integer> {
+public interface ResourceRepository extends CrudRepository<Resource, Long> {
 }
